+++ tests/setup.ts
@@ -1,3 +1,0 @@
-import { loadNative } from "../bin/bun-native";
-
-await (await loadNative()).setupGlobals();